lấy dữ liệu bàn đồ từ bảng spatial data

+++ src/aspnet-core 7.0/src/KNTC.Application/Denounces/DeleteDenounceHandler.cs
@@ -1,12 +1,6 @@
-﻿using KNTC.Denounces;
-using KNTC.SpatialDatas;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using KNTC.SpatialDatas;
 using System.Threading.Tasks;
 using Volo.Abp.DependencyInjection;
-using Volo.Abp.Domain.Repositories;
 using Volo.Abp.EventBus.Distributed;
 
 namespace KNTC.Denounces;
@@ -17,15 +11,20 @@ public class DeleteDenounceHandler
 {
     private readonly ISpatialDataRepository _spatialDataRepo;
     private readonly SpatialDataManager _spatialDataManager;
+
     public DeleteDenounceHandler(ISpatialDataRepository spatialDataRepo,
         SpatialDataManager spatialDataManager)
     {
         _spatialDataRepo = spatialDataRepo;
         _spatialDataManager = spatialDataManager;
     }
+
     public async Task HandleEventAsync(DeleteDenounceEto eventData)
     {
-        var spatialData = await _spatialDataRepo.GetByIdHoSoAsync(eventData.Id);
-        await _spatialDataRepo.DeleteAsync(spatialData);
+        var spatialData = await _spatialDataRepo.FindByIdHoSoAsync(eventData.Id);
+        if (spatialData != null)
+        {
+            await _spatialDataRepo.DeleteAsync(spatialData);
+        }
     }
 }
